Refactor 1 (#97)
* Refactor 1

* Made all references to streetAddress1
